+++ YogiYoCloneIOS/YogiYoCloneIOS/ViewController/ViewControllers/Byun/SearchVC/Controller/DidSearchVC.swift
@@ -113,31 +113,3 @@ extension DidSearchVC : UITextFieldDelegate{
   }
 }
 
-extension DidSearchVC : UITableViewDataSource {
-  func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-    searchList.count
-  }
-  
-  func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-    let cell = tableView.dequeueReusableCell(withIdentifier: "StoreListCell", for: indexPath) as! StoreListCell
-    
-
-    let search = searchList[indexPath.row].results
-    let searchRetrun = searchMager.retunArray()
-    let searchArr = search?[searchRetrun]
-    
-    let imageurl = URL(string: searchArr?.image ?? "")
-    
-    cell.storeImage.kf.setImage(with: imageurl)
-    cell.storeNameLabel.text = searchArr?.name ?? ""
-    cell.storeRateLabel.text = String(searchArr?.star ?? 0.0)
-    cell.reviewLabel.text = String(searchArr?.reviewCount ?? 0)
-    cell.bestMenuLabel.text = searchArr?.representativeMenus ?? ""
-    cell.estimatedTime.text = searchArr?.deliveryTime
-    cell.deliveryDiscountLabel.text = String(searchArr?.deliveryDiscount ?? 0)
-  
-    return cell
-  }
-  
-
-}
